# Optimize digest.ResourceName.Validate