This was using a regexp just to check the length, and that all the characters were lower case hex. In prod this is using roughly 2% of the CPU of pebble_cache.FindMissing

This speeds up pebble read benchmarks by another 3%-14%:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
                                       │     base     │               simple                │
                                       │    sec/op    │    sec/op     vs base               │
GetMulti/size=1KiB-24                    1.525m ±  2%   1.468m ±  1%   -3.74% (p=0.001 n=7)
GetMulti/size=1MiB-24                    65.23m ±  1%   63.26m ±  2%   -3.02% (p=0.002 n=7)
GetMulti/size=10MiB-24                   908.5m ±  2%   882.0m ±  4%   -2.91% (p=0.017 n=7)
FindMissing/size=1KiB/insert=false-24    351.1µ ±  4%   298.9µ ±  4%  -14.85% (p=0.001 n=7)
FindMissing/size=1MiB/insert=false-24    322.5µ ±  1%   274.5µ ±  1%  -14.90% (p=0.001 n=7)
FindMissing/size=10MiB/insert=false-24   329.8µ ±  1%   280.7µ ±  1%  -14.89% (p=0.001 n=7)
FindMissing/size=1KiB/insert=true-24     370.5µ ±  3%   318.6µ ±  3%  -14.03% (p=0.001 n=7)
FindMissing/size=1MiB/insert=true-24     380.2µ ±  4%   330.7µ ±  2%  -13.01% (p=0.001 n=7)
FindMissing/size=10MiB/insert=true-24    383.9µ ±  4%   338.4µ ±  6%  -11.86% (p=0.001 n=7)
Contains1/size=1KiB-24                   5.777µ ±  2%   5.333µ ±  3%   -7.69% (p=0.001 n=7)
Contains1/size=1MiB-24                   6.390µ ±  1%   5.885µ ±  4%   -7.90% (p=0.001 n=7)
Contains1/size=10MiB-24                  6.418µ ±  3%   5.953µ ±  3%   -7.25% (p=0.001 n=7)
Set/size=1KiB-24                         106.5µ ± 13%   109.3µ ±  6%        ~ (p=1.000 n=7)
Set/size=1MiB-24                         3.743m ± 21%   3.712m ± 27%        ~ (p=0.710 n=7)
Set/size=10MiB-24                        25.42m ± 14%   25.35m ±  5%        ~ (p=0.456 n=7)
geomean                                  597.2µ         550.5µ         -7.82%

                                       │      base       │                simple                 │
                                       │      B/op       │      B/op        vs base              │
GetMulti/size=1KiB-24                    737.6Ki ±    0%   736.5Ki ±    0%  -0.15% (p=0.001 n=7)
GetMulti/size=1MiB-24                    188.2Mi ±    1%   185.7Mi ±    1%  -1.33% (p=0.007 n=7)
GetMulti/size=10MiB-24                   2.105Gi ±    1%   2.067Gi ±    2%  -1.80% (p=0.007 n=7)
FindMissing/size=1KiB/insert=false-24    269.1Ki ±    0%   269.0Ki ±    0%  -0.04% (p=0.001 n=7)
FindMissing/size=1MiB/insert=false-24    269.1Ki ±    0%   269.0Ki ±    0%  -0.03% (p=0.001 n=7)
FindMissing/size=10MiB/insert=false-24   269.1Ki ±    0%   269.0Ki ±    0%  -0.04% (p=0.001 n=7)
FindMissing/size=1KiB/insert=true-24     299.0Ki ±    0%   298.9Ki ±    0%  -0.03% (p=0.001 n=7)
FindMissing/size=1MiB/insert=true-24     784.6Ki ±    1%   724.8Ki ±    2%  -7.61% (p=0.001 n=7)
FindMissing/size=10MiB/insert=true-24    791.2Ki ±    3%   733.2Ki ±    4%  -7.34% (p=0.001 n=7)
Contains1/size=1KiB-24                   3.386Ki ±    0%   3.385Ki ±    0%  -0.03% (p=0.001 n=7)
Contains1/size=1MiB-24                   5.585Ki ±    6%   5.388Ki ±    6%       ~ (p=0.535 n=7)
Contains1/size=10MiB-24                  5.413Ki ±    7%   5.286Ki ±    6%       ~ (p=0.318 n=7)
Set/size=1KiB-24                         9.741Ki ± 1062%   9.731Ki ± 1136%       ~ (p=0.902 n=7)
Set/size=1MiB-24                         4.207Mi ±   22%   4.227Mi ±   27%       ~ (p=0.805 n=7)
Set/size=10MiB-24                        29.38Mi ±   17%   29.98Mi ±    5%       ~ (p=0.165 n=7)
geomean                                  555.1Ki           546.8Ki          -1.50%

                                       │      base      │                 simple                 │
                                       │   allocs/op    │   allocs/op     vs base                │
GetMulti/size=1KiB-24                    10.21k ±    0%   10.21k ±    0%       ~ (p=0.070 n=7)
GetMulti/size=1MiB-24                    1.505M ±    3%   1.461M ±    4%  -2.97% (p=0.011 n=7)
GetMulti/size=10MiB-24                   19.71M ±    2%   19.05M ±    3%  -3.39% (p=0.007 n=7)
FindMissing/size=1KiB/insert=false-24    6.416k ±    0%   6.416k ±    0%       ~ (p=1.000 n=7) ¹
FindMissing/size=1MiB/insert=false-24    6.416k ±    0%   6.416k ±    0%       ~ (p=1.000 n=7) ¹
FindMissing/size=10MiB/insert=false-24   6.416k ±    0%   6.416k ±    0%       ~ (p=1.000 n=7) ¹
FindMissing/size=1KiB/insert=true-24     6.108k ±    0%   6.108k ±    0%       ~ (p=1.000 n=7) ¹
FindMissing/size=1MiB/insert=true-24     14.24k ±    1%   13.24k ±    2%  -6.99% (p=0.001 n=7)
FindMissing/size=10MiB/insert=true-24    14.35k ±    3%   13.38k ±    4%  -6.74% (p=0.001 n=7)
Contains1/size=1KiB-24                    69.00 ±    0%    69.00 ±    0%       ~ (p=1.000 n=7) ¹
Contains1/size=1MiB-24                    105.0 ±    6%    102.0 ±    6%       ~ (p=0.396 n=7)
Contains1/size=10MiB-24                   102.0 ±    7%    100.0 ±    6%       ~ (p=0.366 n=7)
Set/size=1KiB-24                          150.0 ± 1155%    150.0 ± 1235%       ~ (p=0.932 n=7)
Set/size=1MiB-24                         71.67k ±   22%   72.02k ±   27%       ~ (p=0.710 n=7)
Set/size=10MiB-24                        489.4k ±   17%   499.7k ±    4%       ~ (p=0.073 n=7)
geomean                                  9.389k           9.246k          -1.52%
```